Initial checkin
This is the first checkin containing the basic first thoughts
of what this project will eventually implement. This project
is an experiment to hash out the ideas that could be used to
create a single-instruction core that is both simple (effient)
and performant (flexible, high logic utilization, decent IPC).

Wish me luck (thanks).

Signed-off-by: davidkoltak <davidkoltak@gmail.com>